+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Alle nennenswerten Änderungen an diesem Projekt werden hier dokumentiert.
 
+## [1.1] - 2026-06-02
+
+### Hinzugefügt
+- Schreibzugriff auf die Pumpe via DABEsy_SetParameter (automatische Codierung anhand der Geräte-Metadaten)
+- Button "Schreibbare Parameter auflisten" zeigt account-abhängig die änderbaren Parameter mit Wertebereich
+- Ausgewählte Variablen (z.B. Soll-Druck) als bedienbare Standardaktion freigeschaltet
+
 ## [1.0] - 2026-05-31
 
 ### Hinzugefügt